Add operators to key release policy
