+++ src/index.ts
@@ -9,6 +9,6 @@ export * from "./endpoints/keyReleasePolicyEndpoint";
 export * from "./endpoints/IKeyItem";
 export * from "./endpoints/KeyGeneration";
 export * from "./endpoints/TinkKey";
-export * from "./policies/IKeyReleasePolicyProps";
+export * from "./policies/IKeyReleasePolicySnpProps";
 export * from "./utils/Tooling";
 export * from "./utils/ServiceResult";
